+++ internal/serve/httphandler/sep10_handler.go
@@ -2,6 +2,7 @@ package httphandler
 
 import (
 	"encoding/json"
+	"errors"
 	"mime"
 	"net/http"
 
@@ -33,6 +34,11 @@ func (h SEP10Handler) GetChallenge(w http.ResponseWriter, r *http.Request) {
 
 	challenge, err := h.SEP10Service.CreateChallenge(ctx, req)
 	if err != nil {
+		var validationErr *services.ChallengeValidationError
+		if errors.As(err, &validationErr) {
+			httperror.BadRequest(err.Error(), err, nil).Render(w)
+			return
+		}
 		httperror.InternalError(ctx, "Failed to create challenge", err, nil).Render(w)
 		return
 	}